fix: Example code in README.md didn't work
The example code in the README was seemingly using old API (getter
methods, rather than accessing properties directly) and querying a
component version with no reported vulnerabilities.

This patch fixs both issues so that the example code works. It also
adds import statements so that the example can be copy/paste/run.

Signed-off-by: Joshua Lock <joshua.lock@uk.verizon.com>